#3221 -- Add checks to filesystem deletes.

+++ src/Controller/Api/Stations/Files/BatchAction.php
@@ -67,7 +67,9 @@ class BatchAction
                         if ($media instanceof Entity\StationMedia) {
                             // Remove album art and waveforms
                             foreach ($media->getRelatedFilePaths() as $relatedFilePath) {
-                                $fs->delete($relatedFilePath);
+                                if ($fs->has($relatedFilePath)) {
+                                    $fs->delete($relatedFilePath);
+                                }
                             }
 
                             // Clear playlists from media
